+++ packages/example/src/functions/functions.commons.ts
@@ -1,7 +1,5 @@
-import { Types } from '../types'
-import m from '@mondrian-framework/module'
 import { PrismaClient } from '@prisma/client'
 
 export type AuthInfo = { userId: string }
 export type SharedContext = { auth?: AuthInfo; prisma: PrismaClient }
-export default m.functionBuilder<Types, SharedContext>()
+