+++ packages/graphql/src/api.ts
@@ -1,7 +1,7 @@
 import { Functions, Logger } from '@mondrian-framework/module'
 import { GraphQLError, GraphQLErrorOptions } from 'graphql'
 
-export type GraphqlFunctionSpecs = { type: 'query' | 'mutation'; name?: string; inputName?: string }
+export type GraphqlFunctionSpecs = { type: 'query' | 'mutation'; name?: string; inputName?: string; namespace?: string | null }
 
 export type GraphqlApi<F extends Functions> = {
   functions: {