+++ packages/aws-lambda-rest/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @mondrian-framework/aws-lambda-rest
 
+## 1.0.5
+
+### Patch Changes
+
+- add namespaces
+- Updated dependencies
+  - @mondrian-framework/rest@1.0.7
+
 ## 1.0.4
 
 ### Patch Changes