+++ packages/graphql-yoga/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @mondrian-framework/graphql-yoga
 
+## 1.0.8
+
+### Patch Changes
+
+- add namespaces
+- Updated dependencies
+  - @mondrian-framework/graphql@1.0.8
+
 ## 1.0.7
 
 ### Patch Changes