+++ packages/graphql/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @mondrian-framework/graphql
 
+## 1.0.8
+
+### Patch Changes
+
+- add namespaces
+
 ## 1.0.7
 
 ### Patch Changes