+++ packages/rest/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @mondrian-framework/rest
 
+## 1.0.7
+
+### Patch Changes
+
+- add namespaces
+
 ## 1.0.6
 
 ### Patch Changes